adding go-git-provider support for acceptance tests
