+++ .github/workflows/deploy.yaml
@@ -93,6 +93,7 @@ jobs:
     timeout-minutes: 90
     env:
       GO_CACHE_NAME: cache-go-modules
+      GIT_PROVIDER: github
       GITHUB_TOKEN: ${{ secrets.WGE_GITHUB_TOKEN }}
       GITHUB_ORG: ${{ secrets.WGE_GITHUB_ORG }}
       GITHUB_USER: ${{ secrets.WGE_GITHUB_USER }}
@@ -105,6 +106,7 @@ jobs:
       SELENIUM_DEBUG: true
       ACCEPTANCE_TESTS_DATABASE_TYPE: sqlite
       ARTEFACTS_BASE_DIR: /tmp/workspace/test/
+      CHECKPOINT_DISABLE: 1
     steps:
     - id: go-cache-paths
       run: |
