+++ test/acceptance/test/acceptance_suite_test.go
@@ -3,7 +3,6 @@ package acceptance
 import (
 	"fmt"
 	"os"
-	"path"
 	"testing"
 
 	"github.com/onsi/ginkgo"
@@ -24,8 +23,8 @@ func GomegaFail(message string, callerSkip ...int) {
 	}
 
 	// Show management cluster pods etc.
-	_ = ShowItems("")
-	_ = DumpClusterInfo(randID)
+	_ = showItems("")
+	_ = dumpClusterInfo(randID)
 
 	//Pass this down to the default handler for onward processing
 	ginkgo.Fail(message, callerSkip...)
@@ -61,18 +60,7 @@ var _ = BeforeSuite(func() {
 	//Set the suite level defaults
 
 	SetDefaultEventuallyTimeout(ASSERTION_DEFAULT_TIME_OUT) //Things are slow on WKP UI
-
-	SELENIUM_SERVICE_URL = "http://localhost:4444/wd/hub"
-	GITHUB_USER = os.Getenv("GITHUB_USER")
-	GITHUB_PASSWORD = os.Getenv("GITHUB_PASSWORD")
-	GIT_PROVIDER = os.Getenv("GIT_PROVIDER")
-	GITHUB_ORG = os.Getenv("GITHUB_ORG")
-	GITHUB_TOKEN = os.Getenv("GITHUB_TOKEN")
-	CLUSTER_REPOSITORY = os.Getenv("CLUSTER_REPOSITORY")
-	GIT_REPOSITORY_URL = "https://" + path.Join("github.com", GITHUB_ORG, CLUSTER_REPOSITORY)
-
-	DOCKER_IO_USER = os.Getenv("DOCKER_IO_USER")
-	DOCKER_IO_PASSWORD = os.Getenv("DOCKER_IO_PASSWORD")
+	SetupTestEnvironment()                                  // Read OS environment variables and initialize the test environment
 })
 
 var _ = AfterSuite(func() {
